Check that marker is on map before setting the icon
There is a race condition in the case the following situation happens:

1. User updates the icon of a Pin with a BitmapDescriptorFactory.FromView()
2. Immediately after remove the Pin from the map

This results in this library calling SetIcon() on the native marker AFTER it has been removed from the native map. This results in Google Maps throwing a `Java.Lang.IllegalArgumentException: Unmanaged descriptor` exception.

By checking that the marker hasn't been removed before calling SetIcon() we can prevent this exception.